restructure API, add data manipulation/querying
