+++ crates/api/src/schema/subscription.rs
@@ -1,0 +1,27 @@
+use std::pin::Pin;
+
+use futures::{Stream, TryStreamExt};
+use juniper::{graphql_subscription, FieldError, FieldResult};
+
+use crate::schema::context::Context;
+use crate::{error::Result, types::report::Report};
+
+type ReportStream = Pin<Box<dyn Stream<Item = FieldResult<Report>> + Send>>;
+
+pub struct Subscription;
+
+#[graphql_subscription(context = Context)]
+impl Subscription {
+    async fn reports(context: &Context) -> ReportStream {
+        let topic = &context.config().report_topic;
+        let stream = context
+            .subscribe_on_kafka_topic(topic)
+            .await?
+            .try_filter_map(|ev| async move { Ok(ev.payload) })
+            .and_then(
+                |payload| async move { serde_json::from_str(&payload).map_err(FieldError::from) },
+            );
+
+        Box::pin(stream)
+    }
+}
