+++ crates/api/src/types/mod.rs
@@ -1,0 +1,3 @@
+pub mod data;
+pub mod report;
+pub mod schema;
